Fix bug when starting with $-sign
When a text token was running, it wouldn't stop on a $-sign if it wasn't followed with a "[".
This is good.

But when a token was finished and the next char was a $-sign, it would consider it an MFMOpen, even when not followed by a "[".
This is now fixed.